+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1.0
 
+* Add `MemPack` instance for lazy `ByteString`
 * Add `MemPack` instance for `Text`
 * Add `MemPack` instance for `Void`
 * Add `packWithByteArray` and `packWithMutableByteArray`